refactor(mcp/write): #465 extract CLUSTER_D format-style pipeline into write_format.py (STRATEGY_S1)
Mechanical VERBATIM extraction (step 3 of 4) of the STRATEGY_S1 sequence
under parent issue #459. Predecessors: PR #467 (CLUSTER_A — write_detection.py),
PR #469 (CLUSTER_B — write_metrics.py).

SEAM_3 designation: per ho-liaison §A.2 this cluster sits on the literal-zone
preservation pipeline (Strategy A preserve mode). The functions thread NFC
bytes, format_style enum, and baseline spans across the single _emit_with_style
emitter surface called from 3+ sites in WriteTool.execute.

Symbols moved from src/octave_mcp/mcp/write.py to new module
src/octave_mcp/mcp/write_format.py (11 declared + co-migrating support):
- _is_atom_value
- _subtree_has_comment
- _block_compact_eligible
- _block_would_collapse_but_for_arity
- _block_to_compact_assignment
- _compact_pass
- _expand_pass
- _maybe_lift_assignment_to_block
- _apply_format_style
- _to_baseline_bytes
- _emit_with_style

Additional co-migrated symbols (consumed exclusively by the moved set and
its external importers — kept colocated for cohesion, no re-export shim):
- FORMAT_STYLE_VALUES (constant tuple)
- E_INVALID_FORMAT_STYLE, W_COMPACT_REFUSED, E_AST_CYCLE (error codes)
- OctaveASTCycleError (exception class with E_AST_CYCLE code)
- _COMPACT_MAX_PAIRS, _REFUSE_REASON_COMMENT, _REFUSE_REASON_ARITY (internal)

Importer rewires:
- src/octave_mcp/mcp/write.py imports the 6 symbols it still uses internally
  (E_INVALID_FORMAT_STYLE, FORMAT_STYLE_VALUES, W_COMPACT_REFUSED,
  OctaveASTCycleError, _emit_with_style, _to_baseline_bytes) from write_format.
  No re-export shim — other consumers import directly from write_format.
- src/octave_mcp/cli/main.py: OctaveASTCycleError + _emit_with_style +
  _to_baseline_bytes import path moved from .write to .write_format
  (_mark_dirty remains in .write — out of CLUSTER_D scope).
- tests/integration/test_write_preserve_nested.py: 4 in-function imports of
  _to_baseline_bytes redirected to write_format.
- tests/unit/test_format_style.py: top-level imports of the 11 symbols +
  constants redirected to write_format; WriteTool import remains from .write.
  Two cyclic_apply monkeypatches updated to target write_format._apply_format_style
  (mechanical import-path update — the monkeypatched function must live where
  _emit_with_style looks it up; no test-logic changes).
- src/octave_mcp/mcp/validate.py + eject.py: no imports affected.

Cleanup in write.py (transitive, pre-existing imports now orphaned):
- removed `import copy` (only _apply_format_style used it)
- removed Comment from cst import (only used by moved helpers)
- removed normalize_content from lexer import (only used by _to_baseline_bytes)

5 regression fences explicitly verified post-extraction (SEAM_3 contract):
- tests/integration/test_write_preserve_nested.py — PASS
- tests/integration/test_schema_write_idempotency.py — PASS
- tests/unit/test_literal_zones_write.py — PASS
- tests/unit/test_octave_write_noop_invariant.py — PASS
- tests/unit/test_writer_reader_symmetry.py — PASS
All 125 tests in the fence set pass byte-identically. format_style ∈
{preserve, expanded, compact, omitted} byte-shape verified via the
idempotency CI gate in test_schema_write_idempotency.py.

LOC delta:
- src/octave_mcp/mcp/write.py: 3917 → 3407 (−510)
- src/octave_mcp/mcp/write_format.py: new, 537 LOC (module docstring +
  imports + 11 functions + 5 constants + OctaveASTCycleError class)

Quality gates:
- ruff: clean
- black: clean
- mypy: clean (3 source files checked)
- pytest: 3614 passed, 11 skipped, 3 xfailed (matches post-CLUSTER_B baseline,
  zero regressions)

PROD::I1 SYNTACTIC_FIDELITY preserved — no signature change, no body edit,
no rename. Pure structural relocation across the SEAM_3 boundary.

Refs: #459 #465
Predecessors: #463 / PR #467, #464 / PR #469
